Rename ObjectStoreCursor to AsyncCursor
