--- src/cursor.rs
@@ -1,117 +1,0 @@
-use std::io::Cursor;
-use std::ops::Range;
-
-use byteorder::{BigEndian, LittleEndian, ReadBytesExt};
-use bytes::Bytes;
-
-use crate::error::AiocogeoError;
-use crate::AsyncFileReader;
-
-#[derive(Debug, Clone, Copy, Default)]
-pub enum Endianness {
-    #[default]
-    LittleEndian,
-    BigEndian,
-}
-
-/// A wrapper around an [ObjectStore] that provides a seek-oriented interface
-// TODO: in the future add buffering to this
-pub(crate) struct ObjectStoreCursor {
-    reader: Box<dyn AsyncFileReader>,
-    offset: usize,
-    endianness: Endianness,
-}
-
-/// Macro to generate functions to read scalar values from the cursor
-macro_rules! impl_read_byteorder {
-    ($method_name:ident, $typ:ty) => {
-        pub(crate) async fn $method_name(&mut self) -> $typ {
-            let mut buf = Cursor::new(self.read(<$typ>::BITS as usize / 8).await);
-            match self.endianness {
-                Endianness::LittleEndian => buf.$method_name::<LittleEndian>().unwrap(),
-                Endianness::BigEndian => buf.$method_name::<BigEndian>().unwrap(),
-            }
-        }
-    };
-}
-
-impl ObjectStoreCursor {
-    pub(crate) fn new(reader: Box<dyn AsyncFileReader>) -> Self {
-        Self {
-            reader,
-            offset: 0,
-            endianness: Default::default(),
-        }
-    }
-
-    pub(crate) fn set_endianness(&mut self, endianness: Endianness) {
-        self.endianness = endianness;
-    }
-
-    pub(crate) fn into_inner(self) -> Box<dyn AsyncFileReader> {
-        self.reader
-    }
-
-    pub(crate) async fn read(&mut self, length: usize) -> Bytes {
-        let range = self.offset..self.offset + length;
-        self.offset += length;
-        self.reader.get_bytes(range).await.unwrap()
-    }
-
-    /// Read a u8 from the cursor
-    pub(crate) async fn read_u8(&mut self) -> u8 {
-        let buf = self.read(1).await;
-        Cursor::new(buf).read_u8().unwrap()
-    }
-
-    /// Read a i8 from the cursor
-    pub(crate) async fn read_i8(&mut self) -> i8 {
-        let buf = self.read(1).await;
-        Cursor::new(buf).read_i8().unwrap()
-    }
-
-    impl_read_byteorder!(read_u16, u16);
-    impl_read_byteorder!(read_u32, u32);
-    impl_read_byteorder!(read_u64, u64);
-    impl_read_byteorder!(read_i16, i16);
-    impl_read_byteorder!(read_i32, i32);
-    impl_read_byteorder!(read_i64, i64);
-
-    pub(crate) async fn read_f32(&mut self) -> f32 {
-        let mut buf = Cursor::new(self.read(4).await);
-        match self.endianness {
-            Endianness::LittleEndian => buf.read_f32::<LittleEndian>().unwrap(),
-            Endianness::BigEndian => buf.read_f32::<BigEndian>().unwrap(),
-        }
-    }
-
-    pub(crate) async fn read_f64(&mut self) -> f64 {
-        let mut buf = Cursor::new(self.read(8).await);
-        match self.endianness {
-            Endianness::LittleEndian => buf.read_f64::<LittleEndian>().unwrap(),
-            Endianness::BigEndian => buf.read_f64::<BigEndian>().unwrap(),
-        }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn reader(&self) -> &dyn AsyncFileReader {
-        &self.reader
-    }
-
-    pub(crate) async fn get_range(&mut self, range: Range<usize>) -> Result<Bytes, AiocogeoError> {
-        self.reader.get_bytes(range).await
-    }
-
-    /// Advance cursor position by a set amount
-    pub(crate) fn advance(&mut self, amount: usize) {
-        self.offset += amount;
-    }
-
-    pub(crate) fn seek(&mut self, offset: usize) {
-        self.offset = offset;
-    }
-
-    pub(crate) fn position(&self) -> usize {
-        self.offset
-    }
-}
